refactor: Moved url input field to separate file
