+++ lib/src/widgets/url_input_field.dart
@@ -1,0 +1,40 @@
+import 'package:easy_localization/easy_localization.dart';
+import 'package:flutter/material.dart';
+
+import '../../l10n/locale_keys.g.dart';
+import '../recipe_controller.dart';
+
+/// A text input field for the URL.
+///
+/// The input is validated to be a valid URL and to be supported by the app.
+/// See [isUrlSupported] for details.
+class UrlInputField extends TextFormField {
+  /// Creates a new [UrlInputField] object.
+  UrlInputField({super.key, required super.controller})
+      : super(
+          autovalidateMode: AutovalidateMode.onUserInteraction,
+          decoration: InputDecoration(
+            border: const OutlineInputBorder(),
+            labelText: LocaleKeys.url_input_field_label.tr(),
+          ),
+          keyboardType: TextInputType.url,
+          validator: (value) {
+            if (value == null || value.isEmpty) {
+              return null;
+            }
+
+            var url = Uri.tryParse(value);
+
+            var isUrl = url?.hasAbsolutePath ?? false;
+            if (!isUrl) {
+              return LocaleKeys.url_input_field_invalid_url_text.tr();
+            }
+
+            if (url != null && !isUrlSupported(url)) {
+              return LocaleKeys.url_input_field_unsupported_url_text.tr();
+            }
+
+            return null;
+          },
+        );
+}
